done all work in Lab2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,7 +4,7 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float) top / bottom;
         int result = Math.round(quotient);
         return result;
     }
@@ -12,12 +12,12 @@ public class DebugExercise1 {
     public static void main(String[] args) {
         int t = 10;
         int b = 2;
-        int result = divideThenRound(t, b);
+        int result = (int) divideThenRound(t, b);
         System.out.println("round(" + t + "/" + b + ")=" + result);
 
         int t2 = 9;
         int b2 = 4;
-        int result2 = divideThenRound(t2, b2);
+        int result2 = (int) divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
 
         int t3 = 3;
